move timestamp record test to the generic test run storage suite
