+++ python_modules/dagster/dagster_tests/core_tests/storage_tests/utils/run_storage.py
@@ -1,9 +1,10 @@
 import sys
+import tempfile
 from datetime import datetime
 
 import pendulum
 import pytest
-from dagster import seven
+from dagster import job, op, seven
 from dagster.core.definitions import PipelineDefinition
 from dagster.core.errors import (
     DagsterRunAlreadyExists,
@@ -16,7 +17,12 @@ from dagster.core.host_representation import (
     ExternalRepositoryOrigin,
     ManagedGrpcPythonEnvRepositoryLocationOrigin,
 )
+from dagster.core.instance import DagsterInstance, InstanceType
+from dagster.core.launcher.sync_in_memory_run_launcher import SyncInMemoryRunLauncher
+from dagster.core.run_coordinator import DefaultRunCoordinator
 from dagster.core.snap import create_pipeline_snapshot_id
+from dagster.core.storage.event_log import InMemoryEventLogStorage
+from dagster.core.storage.noop_compute_log_manager import NoOpComputeLogManager
 from dagster.core.storage.pipeline_run import (
     DagsterRun,
     JobBucket,
@@ -24,6 +30,7 @@ from dagster.core.storage.pipeline_run import (
     PipelineRunsFilter,
     TagBucket,
 )
+from dagster.core.storage.root import LocalArtifactStorage
 from dagster.core.storage.runs.migration import REQUIRED_DATA_MIGRATIONS
 from dagster.core.storage.runs.sql_run_storage import SqlRunStorage
 from dagster.core.storage.tags import PARENT_RUN_ID_TAG, ROOT_RUN_ID_TAG
@@ -32,6 +39,7 @@ from dagster.core.utils import make_new_run_id
 from dagster.daemon.daemon import SensorDaemon
 from dagster.daemon.types import DaemonHeartbeat
 from dagster.serdes import serialize_pp
+from dagster.seven.compat.pendulum import create_pendulum_time, to_timezone
 
 win_py36 = seven.IS_WINDOWS and sys.version_info[0] == 3 and sys.version_info[1] == 6
 
@@ -1310,3 +1318,41 @@ class TestRunStorage:
         assert runs_by_tag.get("1").run_id == one.run_id
         assert runs_by_tag.get("2").run_id == two.run_id
         assert runs_by_tag.get("3").run_id == three.run_id
+
+    def test_run_record_timestamps(self, storage):
+        assert storage
+
+        self._skip_in_memory(storage)
+
+        @op
+        def a():
+            pass
+
+        @job
+        def my_job():
+            a()
+
+        with tempfile.TemporaryDirectory() as temp_dir:
+            instance = DagsterInstance(
+                instance_type=InstanceType.EPHEMERAL,
+                local_artifact_storage=LocalArtifactStorage(temp_dir),
+                run_storage=storage,
+                event_storage=InMemoryEventLogStorage(),
+                compute_log_manager=NoOpComputeLogManager(),
+                run_coordinator=DefaultRunCoordinator(),
+                run_launcher=SyncInMemoryRunLauncher(),
+            )
+
+            freeze_datetime = to_timezone(
+                create_pendulum_time(2019, 11, 2, 0, 0, 0, tz="US/Central"), "US/Pacific"
+            )
+
+            with pendulum.test(freeze_datetime):
+                result = my_job.execute_in_process(instance=instance)
+                records = instance.get_run_records(
+                    filters=PipelineRunsFilter(run_ids=[result.run_id])
+                )
+                assert len(records) == 1
+                record = records[0]
+                assert record.start_time == 1572670800.0
+                assert record.end_time == 1572670800.0
